+++ changelog.md
@@ -77,3 +77,7 @@
 * 3.3.5
   * Fix issue #252 -- Kafka 0.11 and 1.0 have more strict validations on compressed batch wrapper message.
     Changed kafka_protocol 1.1.2 has the wrapper timestamp and offsets fixed.
+* 3.4.0
+  * Added `{prefetch, {bytes, Limit}}' consumer config, old `{prefetch_count, N}' config is still supported
+    but converted to `{prefetch, {count, Limit}}' under the hood
+